services/python-parser/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-docx==1.1.2
pdfminer.six==20240706

# LLM + LangChain
langchain==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.8

# Embeddings
sentence-transformers==3.3.1

# Vector DB
qdrant-client==1.12.1

# Async HTTP
httpx==0.27.2
